近くのアイテムを探索する

+++ 3D_Base/Source/GameObject/SpriteObject/Shadow/ShadowManager/CShadowManager.cpp
@@ -64,12 +64,12 @@ void CShadowManager::Update(CPlayerManager* players, ItemManager* item)
 		}
 	}
 
-	for (int i = Player_Max ; i < item->GetItemVectorNum()+ Player_Max;i++)
+	for (size_t i = Player_Max ; i < item->GetItemVectorNum()+ Player_Max;i++)
 	{ 
-		m_pShadow[i]->Update(item->GetItemPos(i - Player_Max));
+		m_pShadow[i]->Update(item->GetItemPos(static_cast<int>(i - Player_Max)));
 	}
 
-	for (int i = Player_Max + item->GetItemVectorNum();i < Shadow_Max;i++)
+	for (size_t i = Player_Max + item->GetItemVectorNum();i < Shadow_Max;i++)
 	{
 		m_pShadow[i]->SetPosition(0, -10.0f, 0);
 	}
